rename testdata and clean dependencies
