+++ packages/api-explorer/__tests__/testdata/anyOf.refReplacer.expected.json
