+++ packages/api-explorer/__tests__/testdata/circular-on-root.json
