+++ packages/api-explorer/__tests__/testdata/config-root-ref-and-nested.json
