+++ packages/api-explorer/__tests__/testdata/json-editor-bug-max-stack.json
