+++ packages/api-explorer/__tests__/testdata/simple2.refReplacer.expected.json
