+++ packages/api-explorer/__tests__/testdata/simple3.refReplacer.expected.json
